Added last elements

+++ JourneyEngine/src/Journey/Core/Random.cpp
@@ -13,7 +13,7 @@ float Random::genFloat()
 {
 	constexpr uint32_t C_MAX_VAL = std::numeric_limits<uint32_t>::max();
 
-	return static_cast<float>(m_distribution(m_randomEngine)) / C_MAX_VAL;
+	return static_cast<float>(m_distribution(m_randomEngine)) / static_cast<float>(C_MAX_VAL);
 }
 
 }
